Fixed resolution bug

+++ src/Constants.java
@@ -9,7 +9,7 @@ public final class Constants {
     public static final String APP_NAME = "osu!Boost";
     public static final Color NOTE = new Color(53,200, 255); // Light blue
     public static final Color NOTE_HOLD = new Color(0,255, 0); // Green
-    public static final int FULL_HD = 1920;
+    public static final double FULL_HD = 1920;
 
     // GUI
     public static final String TITLE_FONT = "Tahoma";
